discord/requirements.txt
#For gemini.py

# discord.py
# python-dotenv
# google-generativeai

#For combinedversion.py
#discord.py
#requests
# google-generativeai
# python-dotenv
#for new
#matplotlib

#v12
discord.py
requests
python-dotenv